requirements/development.txt
-r base.txt
factory-boy==3.2.0
ipdb==0.13.4
pytest==6.2.1
pytest-cov==2.6.1
pytest-django==4.1.0
pytest-sugar==0.9.4
pylint==2.6.0
pylint-django==2.4.2
black==20.8b1
isort==5.7.0
bump2version==1.0.1
djangorestframework==3.12.4